print message when io get_state
